requirements.txt
# Core NeMo and ASR dependencies
nemo_toolkit[asr]>=2.5.0

# PyTorch and Lightning
torch>=2.0.0
pytorch-lightning>=2.0.0

# Configuration and utilities
omegaconf>=2.3.0
hydra-core>=1.3.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
sox>=1.4.1

# Data processing
text-unidecode>=1.3
matplotlib>=3.3.2

# HuggingFace integration
huggingface-hub>=0.23.2
transformers>=4.30.0

# Experiment tracking (optional)
wandb>=0.15.0

# Modal for GPU training (optional)
modal>=0.60.0

# Development and testing
jupyter>=1.0.0
ipywidgets>=8.0.0
pytest>=7.0.0